refactor(server): external-role.ts を role.ts にリネーム
discord/external-role.ts → discord/role.ts
github/external-role.ts → github/role.ts
ディレクトリ名で external であることは自明なため。
